pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepirtools"
version = "1.1.21"
description = "Deep learning-based estimation and inference for item response theory models."
dependencies = [
    'torch',
    'pyro-ppl',
    'numpy'
]
requires-python = ">=3.6"

[project.urls]
Homepage = "https://github.com/cjurban/deepirtools"